Register and login

+++ HomeAccounting.Domain/DTOs/ViewDTOs/ExchangeRatesViewDTO.cs
@@ -11,7 +11,7 @@ namespace HomeAccounting.Domain.DTOs.ViewDTOs
         public int Id { get; set; }
         public string CurrencyFrom { get; set; }
         public string CurrencyTo { get; set; }
-        public double AmountFrom { get; set; }
-        public double AmountTo { get; set; }
+        public decimal AmountFrom { get; set; }
+        public decimal AmountTo { get; set; }
     }
 }
