+++ HomeAccounting.Domain/DTOs/ViewDTOs/TransactionCategoryViewDTO.cs
@@ -10,7 +10,7 @@ namespace HomeAccounting.Domain.DTOs.ViewDTOs
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double? Constraint { get; set; }
-        public double? ParentCategoryName { get; set; }
+        public decimal? Constraint { get; set; }
+        public int? ParentCategoryName { get; set; }
     }
 }
